Fix warnings in 2d/kernel.hpp
> kernel.hpp:172:21: warning: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Wsign-compare]
> kernel.hpp:143:21: warning: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Wsign-compare]

+++ 2d/include/pcl/2d/impl/kernel.hpp
@@ -140,7 +140,7 @@ pcl::kernel<PointT>::gaussianKernel (pcl::PointCloud<PointT> &kernel)
   }
 
   // Normalizing the kernel
-  for (int i = 0; i < kernel.size (); ++i)
+  for (size_t i = 0; i < kernel.size (); ++i)
     kernel[i].intensity /= sum;
 }
 
@@ -169,7 +169,7 @@ pcl::kernel<PointT>::loGKernel (pcl::PointCloud<PointT> &kernel)
   }
 
   // Normalizing the kernel
-  for (int i = 0; i < kernel.size (); ++i)
+  for (size_t i = 0; i < kernel.size (); ++i)
     kernel[i].intensity /= sum;
 }
